requirements.txt
aiohttp==2.1.0
aioodbc==0.0.3
arrow==0.8.0
async-timeout==1.2.1
binaryornot==0.4.0
chardet==2.3.0
coloredlogs==5.2
cookiecutter==1.4.0
humanfriendly==2.3.2
jinja2-time==0.2.0
multidict==2.1.4
poyo==0.4.0
pyodbc==3.0.10
pytz==2017.2
PyYAML==3.12
rethinkdb==2.3.0.post6
telepot==12.0
urllib3==1.20
whichcraft==0.4.0
yarl==0.10.2
